backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
websockets==13.1
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.51.0